backend/requirements.txt
# Core dependencies
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
asyncpg==0.29.0
redis==5.1.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9
httpx==0.27.2

# Vector DB and embeddings
qdrant-client==1.11.3
sentence-transformers==3.1.1

# LLM
google-generativeai==0.8.1

# Testing
hypothesis==6.112.1
pytest==8.3.3
pytest-asyncio==0.24.0

# Voice features (optional - install separately if needed)
# pip install openai-whisper
# Requires: pip install setuptools wheel
